adding insert_data with validation and add validation for drop_table

+++ drop_table.sh
@@ -1,10 +1,13 @@
 #! /user/bin/bash
 
 read -p "enter tabel name to remove " rm_table
-
 cd Databases/$db_name
-rm $rm_table
-rm .$rm_table.meta_data
+
+if [[ -f $rm_table ]]; then
+    rm $rm_table
+    rm .$rm_table.meta_data
+else
+    echo "Table '$rm_table' does not exist."
+fi
 
 cd ..; cd ..
-echo "Table '$rm_table' has been removed successfully."
